mod: organizing modal cpnts
